# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/clinics": {
      "get": {
        "operationId": "ClinicsController_getDirectory",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "telehealth",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "serviceCode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zipCode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "rating",
                "name",
                "reviews",
                "distance"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicDirectoryResponseDto"
                }
              }
            }
          }
        },
        "summary": "Public, paginated clinic directory",
        "tags": [
          "clinics"
        ]
      }
    },
    "/clinics/{slug}": {
      "get": {
        "operationId": "ClinicsController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Clinic URL slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "No active clinic with billing current for that slug"
          }
        },
        "summary": "Public single-clinic profile by slug",
        "tags": [
          "clinics"
        ]
      }
    },
...
    "/patients/me": {
      "get": {
        "operationId": "PatientsController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get the authenticated patient profile",
        "tags": [
          "patients"
        ]
      },
      "put": {
        "operationId": "PatientsController_putMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Update the authenticated patient profile",
        "tags": [
          "patients"
        ]
      },
      "delete": {
        "operationId": "PatientsController_deleteMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Soft-delete the authenticated patient account and end the session",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/me/leads": {
      "get": {
        "operationId": "PatientsController_getMyLeadsHandler",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get the authenticated patient lead history",
        "tags": [
          "patients"
        ]
      }
    },
...
    "/admin/clinics/{id}": {
      "get": {
        "operationId": "AdminClinicsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get a single clinic by id (admin only)",
        "tags": [
          "Admin — Clinics"
        ]
      },
      "put": {
        "operationId": "AdminClinicsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClinicDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Partially update a clinic (admin only)",
        "tags": [
          "Admin — Clinics"
        ]
      }
    },
    "/admin/clinics/{id}/billing": {
      "get": {
        "operationId": "AdminClinicsController_getBilling",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get a clinic's billing info and invoices (admin only)",
        "tags": [
          "Admin — Clinics"
        ]
      }
    },
...
    "/clinic/portal/billing": {
      "get": {
        "operationId": "ClinicBillingController_getBilling",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get the authenticated clinic billing profile and fee estimate",
        "tags": [
          "Clinic Portal — Billing"
        ]
      },
      "put": {
        "operationId": "ClinicBillingController_updateBilling",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClinicBillingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Upsert the authenticated clinic billing profile; a changed billing email is synced to Stripe best-effort",
        "tags": [
          "Clinic Portal — Billing"
        ]
      }
    },
    "/clinic/portal/payment-method": {
      "get": {
        "operationId": "ClinicBillingController_getPaymentMethodInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentMethodDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get the clinic's default payment method, read live from Stripe; null if none",
        "tags": [
          "Clinic Portal — Billing"
        ]
      },
      "post": {
        "operationId": "ClinicBillingController_savePaymentMethodInfo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavePaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Attach a payment method as the clinic's Stripe default",
        "tags": [
          "Clinic Portal — Billing"
        ]
      },
      "delete": {
        "operationId": "ClinicBillingController_removePaymentMethodInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Detach the clinic's default payment method (best-effort)",
        "tags": [
          "Clinic Portal — Billing"
        ]
      }
    },
    "/clinic/portal/subscription/cancel": {
      "post": {
        "operationId": "ClinicBillingController_cancelSubscriptionInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelSubscriptionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Cancel the clinic subscription; stays active through the end of the current month",
        "tags": [
          "Clinic Portal — Billing"
        ]
      }
    },
    "/clinic/portal/invoices": {
      "get": {
        "operationId": "ClinicBillingController_listInvoices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClinicInvoiceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get the authenticated clinic's own invoices, most recent first",
        "tags": [
          "Clinic Portal — Billing"
        ]
      }
    },
...
  },
  "info": {
    "title": "MedAlign Backend API",
    "description": "MedAlign backend API. Authentication uses an HttpOnly `access_token` cookie set on POST /auth/2fa/verify. All non-GET requests require an `x-csrf-token` header matching the value returned by the CSRF middleware (CSRF is skipped when NODE_ENV=development).\n\n## Local test seed data (run `pnpm seed:pj` against your local DB)\n\n**Clinic portal logins** (password `SeedClinic1!` for both):\n- `clinic-vitality@medalign-seed.example.com` — clinic `vitality-hormone-nyc`\n- `clinic-apex@medalign-seed.example.com` — clinic `apex-peptide-telehealth`\n\n**Superadmin login** (for `/admin/clinics*` and `/admin/patients*`): `superadmin@medalign-seed.example.com` / `SeedAdmin1!`\n\n**Soft-deleted patient** (locked out of sign-in; use it to exercise the admin patient deleted-state and lockout paths): `patient-deleted@medalign-seed.example.com` / `SeedPatient1!`\n\n**2FA:** in `NODE_ENV=development` the 6-digit code is printed to the server logs (LoggingEmailSender) instead of being emailed. Read it from the console and POST it to `/auth/2fa/verify` to obtain the session cookie.\n\nFor patient-flow testing, sign up a new user via POST /auth/signup, then submit an assessment (POST /assessments) to get a `sessionId` + `claimToken`.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "ClinicDirectoryResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClinicDirectoryItemDto"
            }
          },
          "totalCount": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize",
          "totalPages"
        ]
      },
      "ClinicProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "stateCode": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "rating": {
            "type": "number"
          },
          "reviewCount": {
            "type": "number"
          },
          "about": {
            "type": "string"
          },
          "differentiators": {
            "type": "string"
          },
          "services": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allServices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "waitTime": {
            "type": "string"
          },
          "telehealth": {
            "type": "boolean"
          },
          "offersLabWork": {
            "type": "boolean"
          },
          "websiteUrl": {
            "type": "string",
            "nullable": true
          },
          "consultationFee": {
            "type": "string"
          },
          "monthlyProgram": {
            "type": "string"
          },
          "financing": {
            "type": "boolean"
          },
          "insurance": {
            "type": "boolean"
          },
          "providerName": {
            "type": "string"
          },
          "credentials": {
            "type": "string"
          },
          "photoCount": {
            "type": "number"
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "photoUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "category",
          "specialty",
          "location",
          "city",
          "stateCode",
          "zipCode",
          "rating",
          "reviewCount",
          "about",
          "differentiators",
          "services",
          "allServices",
          "waitTime",
          "telehealth",
          "offersLabWork",
          "websiteUrl",
          "consultationFee",
          "monthlyProgram",
          "financing",
          "insurance",
          "providerName",
          "credentials",
          "photoCount",
          "logoUrl",
          "photoUrls"
        ]
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "dob": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "UpdateClinicDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Horizon Hormone Health",
            "description": "Clinic name"
          },
          "status": {
            "type": "string",
            "example": "active",
            "description": "One of active, paused, suspended, inactive. An unrecognized value is ignored."
          },
          "about": {
            "type": "string",
            "description": "About the clinic"
          },
          "differentiators": {
            "type": "string",
            "description": "Clinic differentiators"
          },
          "providerName": {
            "type": "string",
            "example": "Dr. Elena Voss, MD",
            "description": "Primary provider name"
          },
          "credentials": {
            "type": "string",
            "description": "Provider credentials"
          },
          "websiteUrl": {
            "type": "string",
            "example": "https://example.com",
            "description": "Clinic website URL"
          },
          "city": {
            "type": "string",
            "example": "Austin",
            "description": "City"
          },
          "stateCode": {
            "type": "string",
            "example": "TX",
            "description": "Two-letter state code"
          },
          "zipCode": {
            "type": "string",
            "example": "78701",
            "description": "ZIP code"
          },
          "telehealth": {
            "type": "boolean",
            "description": "Whether telehealth is available"
          },
          "offersLabWork": {
            "type": "boolean",
            "description": "Whether the clinic offers lab work"
          },
          "financing": {
            "type": "boolean",
            "description": "Whether financing is available"
          },
          "insurance": {
            "type": "boolean",
            "description": "Whether insurance is accepted"
          },
          "insuranceNotes": {
            "type": "string",
            "description": "Insurance notes"
          },
          "consultationFee": {
            "type": "string",
            "example": "$100-$200",
            "description": "Consultation fee band"
          },
          "monthlyProgram": {
            "type": "string",
            "example": "$200-$400",
            "description": "Monthly program fee band"
          },
          "waitTime": {
            "type": "string",
            "example": "1-2 weeks",
            "description": "Typical new patient wait time"
          },
          "rating": {
            "type": "number",
            "example": 4.5,
            "description": "Rating out of 5"
          },
          "reviewCount": {
            "type": "number",
            "example": 120,
            "description": "Review count"
          },
          "isListedInDirectory": {
            "type": "boolean",
            "description": "Whether the clinic is listed in the public directory"
          },
          "logoUrl": {
            "type": "string",
            "example": "https://storage.example.com/clinics/abc/logo.png",
            "description": "Public URL of the clinic logo"
          }
        }
      },
...
      "UpdateClinicBillingDto": {
        "type": "object",
        "properties": {
          "billingEmail": {
            "type": "string",
            "description": "Billing contact email. Synced to Stripe when changed."
          },
          "billingContactName": {
            "type": "string",
            "description": "Billing contact name"
          },
          "addressLine1": {
            "type": "string",
            "description": "Billing address line 1"
          },
          "addressLine2": {
            "type": "string",
            "description": "Billing address line 2"
          },
          "city": {
            "type": "string",
            "description": "Billing city"
          },
          "stateCode": {
            "type": "string",
            "description": "Billing two-letter state code"
          },
          "zipCode": {
            "type": "string",
            "description": "Billing ZIP code"
          },
          "taxId": {
            "type": "string",
            "description": "Tax ID / EIN"
          }
        }
      },
      "PaymentMethodDto": {
        "type": "object",
        "properties": {
          "brand": {
            "type": "string"
          },
          "last4": {
            "type": "string"
          },
          "exp_month": {
            "type": "number"
          },
          "exp_year": {
            "type": "number"
          },
          "stripePaymentMethodId": {
            "type": "string"
          }
        },
        "required": [
          "brand",
          "last4",
          "exp_month",
          "exp_year",
          "stripePaymentMethodId"
        ]
      },
      "SavePaymentMethodDto": {
        "type": "object",
        "properties": {
          "paymentMethodId": {
            "type": "string",
            "description": "Stripe PaymentMethod id to attach as the clinic default"
          }
        },
        "required": [
          "paymentMethodId"
        ]
      },
      "CancelSubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "cancelledAt": {
            "type": "string",
            "description": "ISO-8601 timestamp of when the cancellation was recorded"
          },
          "activeThrough": {
            "type": "string",
            "description": "ISO-8601 timestamp for the last day of the current month, 23:59:59 UTC"
          }
        },
        "required": [
          "cancelledAt",
          "activeThrough"
        ]
      },
      "ClinicInvoiceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "period_start": {
            "type": "string"
          },
          "period_end": {
            "type": "string"
          },
          "lead_count": {
            "type": "number"
          },
          "price_per_lead": {
            "type": "number"
          },
          "platform_fee": {
            "type": "number"
          },
          "total_amount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "description": "draft | open | paid | overdue | void"
          },
          "invoice_url": {
            "type": "object",
            "nullable": true
          },
          "pdf_url": {
            "type": "object",
            "nullable": true
          },
          "due_date": {
            "type": "object",
            "nullable": true
          },
          "paid_at": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "period_start",
          "period_end",
          "lead_count",
          "price_per_lead",
          "platform_fee",
          "total_amount",
          "status"
        ]
      },
...
    }
  }
}